+++ internal/model/flag_priv_test.go
@@ -202,143 +202,3 @@ func TestFlag_isInPercentage(t *testing.T) {
 		})
 	}
 }
-
-func TestFlag_value(t *testing.T) {
-	type fields struct {
-		Disable    bool
-		Rule       string
-		Percentage float64
-		True       interface{}
-		False      interface{}
-		Default    interface{}
-	}
-	type args struct {
-		flagName string
-		user     ffuser.User
-	}
-	type want struct {
-		value         interface{}
-		variationType VariationType
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		want   want
-	}{
-		{
-			name: "Rule disable get default value",
-			fields: fields{
-				Disable: true,
-				True:    "true",
-				False:   "false",
-				Default: "default",
-			},
-			args: args{
-				flagName: "test_689483",
-				user:     ffuser.NewUser("test_689483"),
-			},
-			want: want{
-				value:         "default",
-				variationType: VariationDefault,
-			},
-		},
-		{
-			name: "Get true value if rule pass",
-			fields: fields{
-				True:       "true",
-				False:      "false",
-				Default:    "default",
-				Rule:       "key == \"user66\"",
-				Percentage: 10,
-			},
-			args: args{
-				flagName: "test-flag",
-				user:     ffuser.NewUserBuilder("user66").AddCustom("name", "john").Build(), // combined hash is 9
-			},
-			want: want{
-				value:         "true",
-				variationType: VariationTrue,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := &Flag{
-				Disable:    tt.fields.Disable,
-				Rule:       tt.fields.Rule,
-				Percentage: tt.fields.Percentage,
-				True:       tt.fields.True,
-				False:      tt.fields.False,
-				Default:    tt.fields.Default,
-			}
-
-			got, variationType := f.Value(tt.args.flagName, tt.args.user)
-			assert.Equal(t, tt.want.value, got)
-			assert.Equal(t, tt.want.variationType, variationType)
-		})
-	}
-}
-
-func TestFlag_String(t *testing.T) {
-	type fields struct {
-		Disable    bool
-		Rule       string
-		Percentage float64
-		True       interface{}
-		False      interface{}
-		Default    interface{}
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		want   string
-	}{
-		{
-			name: "All fields",
-			fields: fields{
-				Disable:    false,
-				Rule:       "key eq \"toto\"",
-				Percentage: 10,
-				True:       true,
-				False:      false,
-				Default:    false,
-			},
-			want: "percentage=10%, rule=\"key eq \"toto\"\", true=\"true\", false=\"false\", true=\"false\", disable=\"false\"",
-		},
-		{
-			name: "No rule",
-			fields: fields{
-				Disable:    false,
-				Percentage: 10,
-				True:       true,
-				False:      false,
-				Default:    false,
-			},
-			want: "percentage=10%, true=\"true\", false=\"false\", true=\"false\", disable=\"false\"",
-		},
-		{
-			name: "Default values",
-			fields: fields{
-				True:    true,
-				False:   false,
-				Default: false,
-			},
-			want: "percentage=0%, true=\"true\", false=\"false\", true=\"false\", disable=\"false\"",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := Flag{
-				Disable:    tt.fields.Disable,
-				Rule:       tt.fields.Rule,
-				Percentage: tt.fields.Percentage,
-				True:       tt.fields.True,
-				False:      tt.fields.False,
-				Default:    tt.fields.Default,
-			}
-			got := f.String()
-			assert.Equal(t, tt.want, got, "String() = %v, want %v", got, tt.want)
-		})
-	}
-}
